push2

+++ ft_strdup.c
@@ -3,17 +3,17 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strdup.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: momahrou <momahrou@student.42.fr>          +#+  +:+       +#+        */
+/*   By: medali <medali@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/10 10:45:57 by momahrou          #+#    #+#             */
-/*   Updated: 2025/11/10 10:56:36 by momahrou         ###   ########.fr       */
+/*   Updated: 2025/11/10 13:43:58 by medali           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stddef.h>
 #include <stdlib.h>
 
-int		ft_strlen(char *string);
+size_t	ft_strlen(char *string);
 void	*ft_memcpy(void *dest, void *src, size_t n);
 
 char	*ft_strdup(char *source)
